Move tasks to backdrop.write
Tasks depend on backdrop.write's config. This is a temporary fix for
config issues caused by having backdrop.write and backdrop.transformers
depend on separate config. This will be refactored when the worker
codebase is extracted from backdrop.
